+++ MeusPilas/app/src/main/java/josealencar/com/br/meuspilas/model/VariableAccount.java
@@ -4,13 +4,13 @@ package josealencar.com.br.meuspilas.model;
  * Created by José on 26/05/2015.
  */
 public class VariableAccount {
-    private int idUser;
+    private long idUser;
     private double valueAccount;
     private int dayPayment;
     private String typeAccount;
     private int amountOfInstallments;
 
-    public VariableAccount(int idUser, double valueAccount, int dayPayment, String typeAccount, int amountOfInstallments) {
+    public VariableAccount(long idUser, double valueAccount, int dayPayment, String typeAccount, int amountOfInstallments) {
         this.idUser = idUser;
         this.valueAccount = valueAccount;
         this.dayPayment = dayPayment;
@@ -18,7 +18,7 @@ public class VariableAccount {
         this.amountOfInstallments = amountOfInstallments;
     }
 
-    public int getIdUser() {
+    public long getIdUser() {
         return this.idUser;
     }
